Finishes StatePanel Context Menus

+++ src/main/java/com/ksatstuttgart/usoc/gui/panel/StatePanel.java
@@ -11,6 +11,7 @@ import com.ksatstuttgart.usoc.gui.setup.configuration.entity.data.VarDTO;
 import javafx.application.Platform;
 import javafx.geometry.HPos;
 import javafx.geometry.Insets;
+import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Priority;
@@ -34,6 +35,11 @@ public class StatePanel extends ScrollPane {
     private StatePaneProperties properties;
 
     /**
+     * Vertical Box that holds all Segments
+     */
+    private VBox stateBox;
+
+    /**
      * Creates an instance of the State Panel
      */
     public StatePanel() {
@@ -48,9 +54,8 @@ public class StatePanel extends ScrollPane {
      */
     private void prepareComponents() {
         // All Segment Panes should be inside stateBox
-        VBox stateBox = new VBox();
+        stateBox = new VBox();
         stateBox.setSpacing(5);
-
 
         // Adds each segment to statePane
         for (Segment segment : properties.getSegments()) {
@@ -59,10 +64,71 @@ public class StatePanel extends ScrollPane {
             stateBox.getChildren().add(segmentPane);
         }
 
+        prepareContextMenu();
+
         setMinWidth(200);
         setFitToWidth(true);
         setFitToHeight(true);
         setContent(stateBox);
+    }
+
+    private void prepareContextMenu() {
+        ContextMenu stateBoxContextMenu = new ContextMenu();
+
+        MenuItem addSegmentMenuItem = new MenuItem("Add Segment");
+        addSegmentMenuItem.setOnAction(onAction -> {
+            stateBoxContextMenu.hide();
+
+            TextInputDialog dialog = new TextInputDialog();
+            dialog.setTitle("New Segment");
+            dialog.setHeaderText("New Segment Name");
+
+            Optional<String> result = dialog.showAndWait();
+
+            result.ifPresent(newName -> {
+                Segment newSegment = new Segment();
+                newSegment.setName(newName);
+
+                properties.getSegments().add(newSegment);
+                stateBox.getChildren().add(new SegmentPane(newSegment));
+            });
+
+            onAction.consume();
+        });
+
+        MenuItem expandMenuItem = new MenuItem("Expand all");
+        expandMenuItem.setOnAction(onAction -> {
+            stateBoxContextMenu.hide();
+
+            for (Node node :
+                    stateBox.getChildren()) {
+                ((SegmentPane) node).setExpanded(true);
+            }
+
+            onAction.consume();
+        });
+
+        MenuItem collapseMenuItem = new MenuItem("Collapse all");
+        collapseMenuItem.setOnAction(onAction -> {
+            stateBoxContextMenu.hide();
+
+            for (Node node :
+                    stateBox.getChildren()) {
+                ((SegmentPane) node).setExpanded(false);
+            }
+
+            onAction.consume();
+        });
+
+        stateBoxContextMenu.getItems().addAll(addSegmentMenuItem, new SeparatorMenuItem(),
+                expandMenuItem, collapseMenuItem);
+
+        stateBoxContextMenu.setImpl_showRelativeToWindow(true);
+
+        stateBox.setOnContextMenuRequested(contextMenuEvent -> {
+            stateBoxContextMenu.show(stateBox,
+                    contextMenuEvent.getScreenX(), contextMenuEvent.getScreenY());
+        });
     }
 
     /**
@@ -95,6 +161,7 @@ public class StatePanel extends ScrollPane {
 
         /**
          * Creates a new SegmentPane
+         *
          * @param segment associated segment
          */
         private SegmentPane(Segment segment) {
@@ -136,8 +203,17 @@ public class StatePanel extends ScrollPane {
                 // Create a new State on the current Segment
                 State newState = new State();
 
-                segment.getStates().add(newState);
-                addState(newState);
+                TextInputDialog dialog = new TextInputDialog();
+                dialog.setTitle("New State");
+                dialog.setHeaderText("State Keyword");
+
+                Optional<String> result = dialog.showAndWait();
+
+                result.ifPresent(newName -> {
+                    newState.setKeyword(newName);
+                    segment.getStates().add(newState);
+                    addState(newState);
+                });
             });
             MenuItem renameSegmentItem = new MenuItem("Rename");
             renameSegmentItem.setOnAction(onAction -> {
@@ -147,7 +223,16 @@ public class StatePanel extends ScrollPane {
             MenuItem deleteSegmentItem = new MenuItem("Delete");
             deleteSegmentItem.setOnAction(onAction -> {
                 segmentContextMenu.hide();
-                deleteSegment();
+
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Confirmation Dialog");
+                alert.setHeaderText("Delete Segment");
+                alert.setContentText("Are you sure you want to remove this segment?");
+
+                Optional<ButtonType> result = alert.showAndWait();
+                if (result.get() == ButtonType.OK) {
+                    deleteSegment();
+                }
             });
 
             segmentContextMenu.getItems().addAll(addStateItem, new SeparatorMenuItem(),
@@ -155,16 +240,33 @@ public class StatePanel extends ScrollPane {
             setContextMenu(segmentContextMenu);
         }
 
+        /**
+         * Renames a Segment
+         */
         private void renameSegment() {
-            // TODO
+            TextInputDialog dialog = new TextInputDialog();
+            dialog.setTitle("Rename Segment");
+            dialog.setHeaderText("New Name");
+
+            Optional<String> result = dialog.showAndWait();
+
+            result.ifPresent(newName -> {
+                segment.setName(newName);
+                this.setText(newName);
+            });
         }
 
+        /**
+         * Deletes a given segment from the Panel and from the layout
+         */
         private void deleteSegment() {
-            // TODO
+            properties.getSegments().remove(segment);
+            stateBox.getChildren().remove(this);
         }
 
         /**
          * Adds a state to the grid
+         *
          * @param state state to be added
          */
         private void addState(State state) {
@@ -224,6 +326,7 @@ public class StatePanel extends ScrollPane {
 
         /**
          * Removes a state
+         *
          * @param label label associated to a state
          * @param state state to be removed
          */
@@ -235,6 +338,7 @@ public class StatePanel extends ScrollPane {
 
         /**
          * Creates a Dialog that permits the user to either add or edit a state
+         *
          * @param title title of the dialog. Should either be 'Edit' or 'Add'
          * @param state state to be added/edited.
          * @return new Dialog
